pyproject.toml
[project]
name = "ds-pipelines-viz"
version = "0.1.0"
description = "Dashboard for visualizing Databricks pipeline status"
requires-python = ">=3.11"
dependencies = [
    "cron-descriptor>=2.0.0",
    "croniter>=2.0.0",
    "databricks-sdk>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.0.0",
]
